requirements.txt
fastapi
uvicorn
streamlit
requests
python-multipart
pillow
numpy
pandas
sentence-transformers
torch
faiss-cpu
pytest
pydantic
python-dotenv
